@@ -8,6 +8,7 @@ import android.content.IntentFilter;
 import android.hardware.usb.UsbDevice;
 import android.hardware.usb.UsbDeviceConnection;
 import android.hardware.usb.UsbManager;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
@@ -38,11 +39,10 @@ import com.hoho.android.usbserial.util.SerialInputOutputManager;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.EnumSet;
-import java.util.concurrent.Executors;
 
 public class TerminalFragment extends Fragment implements SerialInputOutputManager.Listener {
 
-    private enum UsbPermission { Unknown, Requested, Granted, Denied };
+    private enum UsbPermission { Unknown, Requested, Granted, Denied }
 
     private static final String INTENT_ACTION_GRANT_USB = BuildConfig.APPLICATION_ID + ".GRANT_USB";
     private static final int WRITE_WAIT_MILLIS = 2000;
@@ -51,8 +51,8 @@ public class TerminalFragment extends Fragment implements SerialInputOutputManag
     private int deviceId, portNum, baudRate;
     private boolean withIoManager;
 
-    private BroadcastReceiver broadcastReceiver;
-    private Handler mainLooper;
+    private final BroadcastReceiver broadcastReceiver;
+    private final Handler mainLooper;
     private TextView receiveText;
     private ControlLines controlLines;
 
@@ -65,7 +65,7 @@ public class TerminalFragment extends Fragment implements SerialInputOutputManag
         broadcastReceiver = new BroadcastReceiver() {
             @Override
             public void onReceive(Context context, Intent intent) {
-                if(intent.getAction().equals(INTENT_ACTION_GRANT_USB)) {
+                if(INTENT_ACTION_GRANT_USB.equals(intent.getAction())) {
                     usbPermission = intent.getBooleanExtra(UsbManager.EXTRA_PERMISSION_GRANTED, false)
                             ? UsbPermission.Granted : UsbPermission.Denied;
                     connect();
@@ -212,7 +212,8 @@ public class TerminalFragment extends Fragment implements SerialInputOutputManag
         UsbDeviceConnection usbConnection = usbManager.openDevice(driver.getDevice());
         if(usbConnection == null && usbPermission == UsbPermission.Unknown && !usbManager.hasPermission(driver.getDevice())) {
             usbPermission = UsbPermission.Requested;
-            PendingIntent usbPermissionIntent = PendingIntent.getBroadcast(getActivity(), 0, new Intent(INTENT_ACTION_GRANT_USB), 0);
+            int flags = Build.VERSION.SDK_INT >= Build.VERSION_CODES.M ? PendingIntent.FLAG_IMMUTABLE : 0;
+            PendingIntent usbPermissionIntent = PendingIntent.getBroadcast(getActivity(), 0, new Intent(INTENT_ACTION_GRANT_USB), flags);
             usbManager.requestPermission(driver.getDevice(), usbPermissionIntent);
             return;
         }
@@ -226,10 +227,10 @@ public class TerminalFragment extends Fragment implements SerialInputOutputManag
 
         try {
             usbSerialPort.open(usbConnection);
-            usbSerialPort.setParameters(baudRate, 8, UsbSerialPort.STOPBITS_1, UsbSerialPort.PARITY_NONE);
+            usbSerialPort.setParameters(baudRate, 8, 1, UsbSerialPort.PARITY_NONE);
             if(withIoManager) {
                 usbIoManager = new SerialInputOutputManager(usbSerialPort, this);
-                Executors.newSingleThreadExecutor().submit(usbIoManager);
+                usbIoManager.start();
             }
             status("connected");
             connected = true;
@@ -243,8 +244,10 @@ public class TerminalFragment extends Fragment implements SerialInputOutputManag
     private void disconnect() {
         connected = false;
         controlLines.stop();
-        if(usbIoManager != null)
+        if(usbIoManager != null) {
+            usbIoManager.setListener(null);
             usbIoManager.stop();
+        }
         usbIoManager = null;
         try {
             usbSerialPort.close();
@@ -258,10 +261,10 @@ public class TerminalFragment extends Fragment implements SerialInputOutputManag
             return;
         }
         try {
-            byte[] data = (str + '\n').getBytes();
+        byte[] data = (str + '\n').getBytes();
             SpannableStringBuilder spn = new SpannableStringBuilder();
             spn.append("send " + data.length + " bytes\n");
-            spn.append(HexDump.dumpHexString(data)+"\n");
+            spn.append(HexDump.dumpHexString(data)).append("\n");
             spn.setSpan(new ForegroundColorSpan(getResources().getColor(R.color.colorSendText)), 0, spn.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
             receiveText.append(spn);
             usbSerialPort.write(data, WRITE_WAIT_MILLIS);
@@ -291,7 +294,7 @@ public class TerminalFragment extends Fragment implements SerialInputOutputManag
         SpannableStringBuilder spn = new SpannableStringBuilder();
         spn.append("receive " + data.length + " bytes\n");
         if(data.length > 0)
-            spn.append(HexDump.dumpHexString(data)+"\n");
+            spn.append(HexDump.dumpHexString(data)).append("\n");
         receiveText.append(spn);
     }
 
@@ -304,8 +307,8 @@ public class TerminalFragment extends Fragment implements SerialInputOutputManag
     class ControlLines {
         private static final int refreshInterval = 200; // msec
 
-        private Runnable runnable;
-        private ToggleButton rtsBtn, ctsBtn, dtrBtn, dsrBtn, cdBtn, riBtn;
+        private final Runnable runnable;
+        private final ToggleButton rtsBtn, ctsBtn, dtrBtn, dsrBtn, cdBtn, riBtn;
 
         ControlLines(View view) {
             runnable = this::run; // w/o explicit Runnable, a new lambda would be created on each postDelayed, which would not be found again by removeCallbacks
