@@ -50,15 +50,29 @@ public final class UsbId {
     public static final int SILABS_CP2108 = 0xea71;
 
     public static final int VENDOR_PROLIFIC = 0x067b;
-    public static final int PROLIFIC_PL2303 = 0x2303;
+    public static final int PROLIFIC_PL2303 = 0x2303;   // device type 01, T, HX
+    public static final int PROLIFIC_PL2303GC = 0x23a3; // device type HXN
+    public static final int PROLIFIC_PL2303GB = 0x23b3; // "
+    public static final int PROLIFIC_PL2303GT = 0x23c3; // "
+    public static final int PROLIFIC_PL2303GL = 0x23d3; // "
+    public static final int PROLIFIC_PL2303GE = 0x23e3; // "
+    public static final int PROLIFIC_PL2303GS = 0x23f3; // "
 
     public static final int VENDOR_QINHENG = 0x1a86;
     public static final int QINHENG_CH340 = 0x7523;
     public static final int QINHENG_CH341A = 0x5523;
+    public static final int QINHENG_CH9102F = 0x55D4;
 
     // at www.linux-usb.org/usb.ids listed for NXP/LPC1768, but all processors supported by ARM mbed DAPLink firmware report these ids
     public static final int VENDOR_ARM = 0x0d28;
     public static final int ARM_MBED = 0x0204;
+
+    public static final int VENDOR_ST = 0x0483;
+    public static final int ST_CDC = 0x5740;
+
+    public static final int VENDOR_RASPBERRY_PI = 0x2e8a;
+    public static final int RASPBERRY_PI_PICO_MICROPYTHON = 0x0005;
+    public static final int RASPBERRY_PI_PICO_SDK = 0x000a;
 
     private UsbId() {
         throw new IllegalAccessError("Non-instantiable class");
