@@ -21,7 +21,7 @@ public interface UsbSerialDriver {
      *
      * @return the device
      */
-    public UsbDevice getDevice();
+    UsbDevice getDevice();
 
     /**
      * Returns all available ports for this device. This list must have at least
@@ -29,5 +29,5 @@ public interface UsbSerialDriver {
      *
      * @return the ports
      */
-    public List<UsbSerialPort> getPorts();
+    List<UsbSerialPort> getPorts();
 }
