@@ -8,10 +8,15 @@ package com.hoho.android.usbserial.driver;
 
 import android.hardware.usb.UsbDevice;
 import android.hardware.usb.UsbDeviceConnection;
+import android.hardware.usb.UsbEndpoint;
 import android.hardware.usb.UsbManager;
+
+import androidx.annotation.IntDef;
 
 import java.io.Closeable;
 import java.io.IOException;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 import java.util.EnumSet;
 
 /**
@@ -22,81 +27,73 @@ import java.util.EnumSet;
 public interface UsbSerialPort extends Closeable {
 
     /** 5 data bits. */
-    public static final int DATABITS_5 = 5;
-
+    int DATABITS_5 = 5;
     /** 6 data bits. */
-    public static final int DATABITS_6 = 6;
-
+    int DATABITS_6 = 6;
     /** 7 data bits. */
-    public static final int DATABITS_7 = 7;
-
+    int DATABITS_7 = 7;
     /** 8 data bits. */
-    public static final int DATABITS_8 = 8;
+    int DATABITS_8 = 8;
 
-    /** No flow control. */
-    public static final int FLOWCONTROL_NONE = 0;
-
-    /** RTS/CTS input flow control. */
-    public static final int FLOWCONTROL_RTSCTS_IN = 1;
-
-    /** RTS/CTS output flow control. */
-    public static final int FLOWCONTROL_RTSCTS_OUT = 2;
-
-    /** XON/XOFF input flow control. */
-    public static final int FLOWCONTROL_XONXOFF_IN = 4;
-
-    /** XON/XOFF output flow control. */
-    public static final int FLOWCONTROL_XONXOFF_OUT = 8;
-
+    /** Values for setParameters(..., parity) */
+    @Retention(RetentionPolicy.SOURCE)
+    @IntDef({PARITY_NONE, PARITY_ODD, PARITY_EVEN, PARITY_MARK, PARITY_SPACE})
+    @interface Parity {}
     /** No parity. */
-    public static final int PARITY_NONE = 0;
-
+    int PARITY_NONE = 0;
     /** Odd parity. */
-    public static final int PARITY_ODD = 1;
-
+    int PARITY_ODD = 1;
     /** Even parity. */
-    public static final int PARITY_EVEN = 2;
-
+    int PARITY_EVEN = 2;
     /** Mark parity. */
-    public static final int PARITY_MARK = 3;
-
+    int PARITY_MARK = 3;
     /** Space parity. */
-    public static final int PARITY_SPACE = 4;
+    int PARITY_SPACE = 4;
 
     /** 1 stop bit. */
-    public static final int STOPBITS_1 = 1;
-
+    int STOPBITS_1 = 1;
     /** 1.5 stop bits. */
-    public static final int STOPBITS_1_5 = 3;
-
+    int STOPBITS_1_5 = 3;
     /** 2 stop bits. */
-    public static final int STOPBITS_2 = 2;
+    int STOPBITS_2 = 2;
 
-    /** values for get[Supported]ControlLines() */
-    public enum ControlLine { RTS, CTS,  DTR, DSR,  CD, RI };
+    /** Values for get[Supported]ControlLines() */
+    enum ControlLine { RTS, CTS,  DTR, DSR,  CD, RI }
 
     /**
      * Returns the driver used by this port.
      */
-    public UsbSerialDriver getDriver();
+    UsbSerialDriver getDriver();
 
     /**
      * Returns the currently-bound USB device.
      */
-    public UsbDevice getDevice();
+    UsbDevice getDevice();
 
     /**
      * Port number within driver.
      */
-    public int getPortNumber();
-    
+    int getPortNumber();
+
+    /**
+     * Returns the write endpoint.
+     * @return write endpoint
+     */
+    UsbEndpoint getWriteEndpoint();
+
+    /**
+     * Returns the read endpoint.
+     * @return read endpoint
+     */
+    UsbEndpoint getReadEndpoint();
+
     /**
      * The serial number of the underlying UsbDeviceConnection, or {@code null}.
      *
      * @return value from {@link UsbDeviceConnection#getSerial()}
      * @throws SecurityException starting with target SDK 29 (Android 10) if permission for USB device is not granted
      */
-    public String getSerial();
+    String getSerial();
 
     /**
      * Opens and initializes the port. Upon success, caller must ensure that
@@ -106,14 +103,14 @@ public interface UsbSerialPort extends Closeable {
      *                   {@link UsbManager#openDevice(android.hardware.usb.UsbDevice)}
      * @throws IOException on error opening or initializing the port.
      */
-    public void open(UsbDeviceConnection connection) throws IOException;
+    void open(UsbDeviceConnection connection) throws IOException;
 
     /**
      * Closes the port and {@link UsbDeviceConnection}
      *
      * @throws IOException on error closing the port.
      */
-    public void close() throws IOException;
+    void close() throws IOException;
 
     /**
      * Reads as many bytes as possible into the destination buffer.
@@ -123,17 +120,18 @@ public interface UsbSerialPort extends Closeable {
      * @return the actual number of bytes read
      * @throws IOException if an error occurred during reading
      */
-    public int read(final byte[] dest, final int timeout) throws IOException;
+    int read(final byte[] dest, final int timeout) throws IOException;
 
     /**
      * Writes as many bytes as possible from the source buffer.
      *
      * @param src the source byte buffer
      * @param timeout the timeout for writing in milliseconds, 0 is infinite
-     * @return the actual number of bytes written
+     * @throws SerialTimeoutException if timeout reached before sending all data.
+     *                                ex.bytesTransferred may contain bytes transferred
      * @throws IOException if an error occurred during writing
      */
-    public int write(final byte[] src, final int timeout) throws IOException;
+    void write(final byte[] src, final int timeout) throws IOException;
 
     /**
      * Sets various serial port parameters.
@@ -147,7 +145,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException on error setting the port parameters
      * @throws UnsupportedOperationException if values are not supported by a specific device
      */
-    public void setParameters(int baudRate, int dataBits, int stopBits, int parity) throws IOException;
+    void setParameters(int baudRate, int dataBits, int stopBits, @Parity int parity) throws IOException;
 
     /**
      * Gets the CD (Carrier Detect) bit from the underlying UART.
@@ -156,7 +154,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during reading
      * @throws UnsupportedOperationException if not supported
      */
-    public boolean getCD() throws IOException;
+    boolean getCD() throws IOException;
 
     /**
      * Gets the CTS (Clear To Send) bit from the underlying UART.
@@ -165,7 +163,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during reading
      * @throws UnsupportedOperationException if not supported
      */
-    public boolean getCTS() throws IOException;
+    boolean getCTS() throws IOException;
 
     /**
      * Gets the DSR (Data Set Ready) bit from the underlying UART.
@@ -174,7 +172,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during reading
      * @throws UnsupportedOperationException if not supported
      */
-    public boolean getDSR() throws IOException;
+    boolean getDSR() throws IOException;
 
     /**
      * Gets the DTR (Data Terminal Ready) bit from the underlying UART.
@@ -183,7 +181,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during reading
      * @throws UnsupportedOperationException if not supported
      */
-    public boolean getDTR() throws IOException;
+    boolean getDTR() throws IOException;
 
     /**
      * Sets the DTR (Data Terminal Ready) bit on the underlying UART, if supported.
@@ -192,7 +190,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during writing
      * @throws UnsupportedOperationException if not supported
      */
-    public void setDTR(boolean value) throws IOException;
+    void setDTR(boolean value) throws IOException;
 
     /**
      * Gets the RI (Ring Indicator) bit from the underlying UART.
@@ -201,7 +199,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during reading
      * @throws UnsupportedOperationException if not supported
      */
-    public boolean getRI() throws IOException;
+    boolean getRI() throws IOException;
 
     /**
      * Gets the RTS (Request To Send) bit from the underlying UART.
@@ -210,7 +208,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during reading
      * @throws UnsupportedOperationException if not supported
      */
-    public boolean getRTS() throws IOException;
+    boolean getRTS() throws IOException;
 
     /**
      * Sets the RTS (Request To Send) bit on the underlying UART, if supported.
@@ -219,7 +217,7 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during writing
      * @throws UnsupportedOperationException if not supported
      */
-    public void setRTS(boolean value) throws IOException;
+    void setRTS(boolean value) throws IOException;
 
     /**
      * Gets all control line values from the underlying UART, if supported.
@@ -228,7 +226,7 @@ public interface UsbSerialPort extends Closeable {
      * @return EnumSet.contains(...) is {@code true} if set, else {@code false}
      * @throws IOException if an error occurred during reading
      */
-    public EnumSet<ControlLine> getControlLines() throws IOException;
+    EnumSet<ControlLine> getControlLines() throws IOException;
 
     /**
      * Gets all control line supported flags.
@@ -236,7 +234,7 @@ public interface UsbSerialPort extends Closeable {
      * @return EnumSet.contains(...) is {@code true} if supported, else {@code false}
      * @throws IOException if an error occurred during reading
      */
-    public EnumSet<ControlLine> getSupportedControlLines() throws IOException;
+    EnumSet<ControlLine> getSupportedControlLines() throws IOException;
 
     /**
      * Purge non-transmitted output data and / or non-read input data.
@@ -246,18 +244,18 @@ public interface UsbSerialPort extends Closeable {
      * @throws IOException if an error occurred during flush
      * @throws UnsupportedOperationException if not supported
      */
-    public void purgeHwBuffers(boolean purgeWriteBuffers, boolean purgeReadBuffers) throws IOException;
+    void purgeHwBuffers(boolean purgeWriteBuffers, boolean purgeReadBuffers) throws IOException;
 
     /**
      * send BREAK condition.
      *
      * @param value set/reset
      */
-    public void setBreak(boolean value) throws IOException;
+    void setBreak(boolean value) throws IOException;
 
     /**
      * Returns the current state of the connection.
      */
-    public boolean isOpen();
+    boolean isOpen();
 
 }
